configured mouse over events on schema tree, changed icon
