+++ client/components/Tree.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import { Treebeard, decorators } from 'react-treebeard';
 import autoQuery from '../../helpers/autoQuery';
-
+import customHeader from '../../helpers/customHeader';
+import customToggle from '../../helpers/customToggle';
 
 
 export default function TreeExample(props) {
@@ -9,29 +10,6 @@ export default function TreeExample(props) {
   const [cursor, setCursor] = useState(false);
   const [data, setData] = useState(tree);
 
-
-  const customHeader = (props) => {
-    let argString = ''
-    const args = props.node.attributes ? Object.entries(props.node.attributes) : '';
-    if(typeof args === 'object') {
-      args.forEach(arg => {
-        argString += `(${arg[0]} : ${arg[1]})`;
-      })
-    }
-    return (
-      <div style={{"display": "inline-block", "verticalAlign": "middle",}}>
-        {`${props.node.name} ${argString}`}
-      </div>
-    )
-  }
-
-  const customToggle = (props) => {
-    return (
-      <div style={{"display": "inline-block", "verticalAlign": "middle", "marginRight" : '5px'}}>
-        <svg width="10" height="10" viewBox="0 0 24 24"><path d="M12 0c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm7 14h-5v5h-4v-5h-5v-4h5v-5h4v5h5v4z" fill="#a8cfc5"/></svg>
-      </div>
-    )
-  }
 
   decorators.Header = customHeader;
   decorators.Toggle = customToggle;
@@ -62,5 +40,5 @@ export default function TreeExample(props) {
         <span className="error-message">{tree.error}</span>
       </>
     );
-  else return <Treebeard data={tree} onToggle={onToggle} decorators={decorators}/>;
+  else return <Treebeard data={tree} onToggle={onToggle} decorators={decorators} animations={false}/>;
 }
