+++ client/stylesheets/tree.scss
@@ -21,7 +21,7 @@
 
 .sidebar-schema,
 .sidebar-schema-description {
-  font-size: 3rem;
+  font-size: 2rem;
   color: lightgray;
   text-align: start;
   border-bottom: 1px solid gray;
